pyproject.toml
[tool.poetry]
name = "api-para-engenheiros-de-dados"
version = "0.1.0"
description = ""
authors = ["gabriel-barata"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.108.0"
uvicorn = "^0.25.0"
faker = "^22.1.0"
python-dotenv = "^1.0.0"
pydantic = {extras = ["email"], version = "^2.5.3"}
sqlalchemy = "^2.0.25"
passlib = "^1.7.4"
bcrypt = "^4.1.2"
python-multipart = "^0.0.6"
alembic = "^1.13.1"
python-jose = "^3.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
